Moving functions/computations into seperate modules for better structure
